Convert to spigot mappings, remove unknown fields
