+++ modules/api/src/main/java/net/countercraft/movecraft/util/teleport/AbstractTeleport.java
